PC-605: move all tests back into test_eligibility.py
no point splitting files now if theyre only one test per scenario
